Refactor several models
One thing I noticed is there was tons of repetitive logic across all of the models for getting the embeddings for entities, in scoring, and in forward/predict. So I refactored a bunch of this into helper functions. I'm sure there's more we could do to make even more of this functionality abstracted but this is a first step. ALSO: I made much nicer names for variables.