Fix bugs in ComplEx model
The regularization factor was not initialized on the model device, but
always on the CPU, which led to crashes when the model was on the GPU.
Furthermore, the regularization factor  was added as a tensor and not as a scalar.
The compute_label_loss method of the base class was called as a private
function instead of a public function.
	geändert:       src/poem/models/unimodal/complex.py