Fix the triples factory instance creator incorrect behaviour
Previously the ids of subject relation pairs were passed to the
dictionary as torch tensors. However, torch tensors are hashable, i.e.
even though the tensors contained the same entity and relation ids they
resolve to different hashes causing the python dictionary to store
identical subject relation pairs as distinct keys.
Torch tensors being hashable is an open issue in the torch library, see
https://github.com/pytorch/pytorch/issues/7733

This issue is fixed by converting the mapped triples tensor to a
numpy array, which is not hashable.
Fixes #172
	geändert:       src/poem/instance_creation_factories/triples_factory.py