Support sklearn evaluators (#193)
* Add wrapper to use sklearn ranking metrics for evaluation
* Remove adjusted MRR
* Add positive mask to evaluator
* Adjust unittest of evaluators to use actual data
* Adjust rank-based evaluator to match overwritten signature
* Ensure that scores are only counted once
* Optimize filtering further
* Make compute_rank_from_scores method public
* Fix the evaluation for the filtered setting
When filtering the scores the true triples are set to "nan" as well, due
to efficiency gains. This causes problems in the following evaluation,
because the true scores are not available anymore. This was fixed by
re-writing the original scores after filtering.
Fixes #198
Fixes #249 partially, only for the filtered setting.
* Update early stopping unit tests affected by code changes