Implement fall back interaction function (#271)
The scoring_h and scoring_t functions are changed to fall back to using scoring_hrt the same way scoring_r does it already. This allows the immediate training of any model under the lcwa assumption with only having the generic scoring_hrt function. Furthermore, this allows to evaluate models that were trained under the owa assumption with the scoring_hrt function as well.