Update predict.py
Fixed indexError where it had flipped the the remainder and quotient. (Error not visible on original code snippet as it had the same amount of entities and relation (making it a square). If say we had 1000 entities and 1 relation (1000 x 1) it would orginially iterate (1 x 1000) getting an index error. This flips these indices.