🪫🔈 Generalize low-rank representation (#1509)
Implement the generalized low-rank representation, where both, the base
and weight are arbitrary representations themselves.

This adds some nice features such as allowing to impose constraints or
regularizes on the weight matrix. An interesting instantiation could be
to use a softmax normalizer on the weight matrix, which gives the
representation the interpretation of a mixture model.


- Extend `LowRankRepresentation` to accept configurable base and weight
representations
- Register `softmax` as normalizer
- Extend `WikidataTextCache` to support retrieving relation information.
- Add example for using the generalized low-rank representation for a
mixture

---------

Co-authored-by: Charles Tapley Hoyt <cthoyt@gmail.com>