🪢🪛 Fix type annotations for entity/relation to ID mapping (#1521)
Previously, it was annotated as `Collection`. However, it is important
the order of the result is the same order as the input, because in many
use cases we want to lookup ids for multiple entities and then use the
ids to e.g. lookup embeddings.

The implementation did already do that, so I just updated the type
annotations.

At the same time, I also allowed to pass `Iterable` to push the
lazy/stream-able interface further down.