💸🏋Loss weights (#1529)
Add sample weights to loss functions.

For many loss functions, this works by selecting none as the aggregation
of the underlying loss implementation and calculating a weighted
average/sum instead of the uniformly weighted average/sum.

This PR does not:

- add support for sample weights to the training loops
- add support for class weights
- some underlying PyTorch loss functions natively support those, e.g.,
[torch.nn.BCEWithLogitsLoss](https://pytorch.org/docs/stable/generated/torch.nn.BCEWithLogitsLoss.html)'s
pos_weight. It would make sense to expose that, but it is a different
feature.
- add support for edge weights in message passing. We already have some
support for them.

Related to https://github.com/pykeen/pykeen/issues/1523

An alternative to https://github.com/pykeen/pykeen/pull/1524

---------

Co-authored-by: Charles Tapley Hoyt <cthoyt@gmail.com>