🔁🧼 Refactor typing in `repeat_if_necessary()` and `Model._get_entity_len()` (#1536)
This function is type annotated to take an optional `num`, but in
practice, it's always given, and in the torch implementation, it doesn't
even accept a None. This PR updates the implementation, the docs, and
makes this function private.

This also led me to identify that `num_entities` within the LCWA
training loop was type annotated as being possible to be none, but in
practice it never is.

---------

Co-authored-by: Max Berrendorf <max.berrendorf@gmail.com>